requirements.txt
asgiref==3.7.2
attrs==23.2.0
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
django==3.2
djangorestframework==3.14.0
exceptiongroup==1.2.0
flatbuffers==23.5.26
idna==3.6
openmeteo_requests==1.2.0
openmeteo_sdk==1.10.0
platformdirs==4.2.0
pytz==2024.1
requests==2.31.0
requests-cache==1.1.1
retry-requests==2.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
url-normalize==1.4.3
urllib3==2.2.0
celery
django-celery-beat
redis
psycopg2-binary